code/requirements.txt
# AWS SDK
boto3>=1.34.0

# GCP SDKs
google-cloud-dataproc>=5.8.0
google-cloud-storage>=2.14.0

# PySpark for distributed computing (includes MLlib for ML)
pyspark>=3.4.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0

# Data handling and downloading
Pillow>=10.0.0
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
pytz>=2024.1

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and logging
pyyaml>=6.0.0
python-dotenv>=1.0.0
